don't forget to update m_block_header in DataInputIterator::+=
After we have found the element the user wants to iterate on,
we need to update the m_block_header so that we properly refer
to it afterwards.

It was not an issue before since += uses m_block_header only to
get the data size, and in the earlier version of typelib all
samples had the same size.